added delete functionality for quiz template

+++ app/src/main/java/com/example/triviaapp/ui/dialogs/savequiztemplate/QuizTemplateDialogState.kt
@@ -21,7 +21,7 @@ sealed class QuizTemplateDialogState {
         ) : Shown()
 
         data class Edit(
-            val quizTemplateId: Long,
+            val quizTemplateId: Int,
             override val name: String? = null,
             override val error: QuizTemplateNameError? = null
         ) : Shown()
